+++ Client.js
@@ -1,0 +1,12 @@
+class Client {
+  // prettier-ignore
+  constructor(id, displayName, fromPhoneNumber, createdAt, updatedAt) {
+    this.id = id
+    this.displayName = displayName
+    this.fromPhoneNumber = fromPhoneNumber
+    this.createdAt = createdAt
+    this.updatedAt = updatedAt
+  }
+}
+
+module.exports = Client