+++ test/integration/BraveAlerterConfiguratorTest/getAlertSessionTest.js
@@ -6,11 +6,13 @@ const { afterEach, beforeEach, describe, it } = require('mocha')
 const { CHATBOT_STATE, ALERT_TYPE, AlertSession } = require('brave-alert-lib')
 const BraveAlerterConfigurator = require('../../../BraveAlerterConfigurator')
 const db = require('../../../db/db')
+const { clientFactory } = require('../../../testingHelpers')
 
 describe('BraveAlerterConfigurator.js integration tests: getAlertSession', () => {
   beforeEach(async () => {
     await db.clearSessions()
     await db.clearLocations()
+    await db.clearClients()
 
     this.expectedChatbotState = CHATBOT_STATE.WAITING_FOR_CATEGORY
     this.expectedIncidentType = 'No One Inside'
@@ -18,6 +20,7 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSession', () =>
     this.expectedLocationPhoneNumber = '+17772225555'
 
     // Insert a location in the DB
+    const client = await clientFactory(db)
     await db.createLocation(
       'LocationId',
       this.expectedLocationPhoneNumber,
@@ -37,6 +40,7 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSession', () =>
       'alertApiKey',
       true,
       false,
+      client.id,
     )
     const locationId = (await db.getLocations())[0].locationid
 
@@ -50,6 +54,7 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSession', () =>
   afterEach(async () => {
     await db.clearSessions()
     await db.clearLocations()
+    await db.clearClients()
   })
 
   it('should create a new AlertSession with expected values from the sessions and locations DB tables', async () => {